Refactored PythonRunner changes after upmerge
